refactor: Store complete URL in variable before making HTTP request
This commit modifies the GetDeviceByImei function to store the complete URL in a variable before making the HTTP request. This change can help with debugging by allowing the complete URL to be logged if an error occurs. The error messages have also been updated to provide more context about the source of the error.