Refactor: Add context with timeout to HTTP request in RequestExecutor
This commit modifies the processURL function in the RequestExecutor class to add a context with a timeout of 10 seconds to the HTTP request. This ensures that if the HTTP request does not complete within 10 seconds, it will be cancelled and an error will be generated. This change helps to avoid long delays in the event of network issues or server problems.